+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Speed up user history load [#449](https://github.com/rokwire/safer-illinois-app/issues/449).
+- Encode health statuses in rules [#452](https://github.com/rokwire/safer-illinois-app/issues/452).
 
 ### Fixed
 - Do not translate symptoms names transmitted to analytics [#447](https://github.com/rokwire/safer-illinois-app/issues/447).